# Conflicts:
#	QuanLyBanHang/src/routes.js
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/controller/BillController.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/controller/CustomerController.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/controller/OrderController.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/dao/GenericDao.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/dao/IOrderDao.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/dao/impl/AbstractDao.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/dao/impl/BillDao.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/dao/impl/OrderDao.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/dto/DashBoardItem.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/dto/OrderDto.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/dto/OrderPageable.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/service/IBillDetailService.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/service/IOrderService.java
#	quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/service/impl/BillDetailService.java
#	quan-ly-ban-hang/src/main/resources/application.properties

+++ quan-ly-ban-hang/src/main/java/com/sapo/quanlybanhang/dto/BillListDto.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 @Data
 public class BillListDto {
-    private Integer totalItem;
+    private Long totalItem;
     private List<BillDto> resultList = new ArrayList<>();
 }
